+++ src/lib/Warp.NET.RenderImpl.Ui/Rendering/Scissors/ScissorScheduler.cs
@@ -2,21 +2,49 @@ namespace Warp.NET.RenderImpl.Ui.Rendering.Scissors;
 
 public static class ScissorScheduler
 {
-	public static Scissor? CurrentScissor { get; private set; }
+	private static readonly List<Scissor> _activeScissors = new();
 
-	/// <summary>
-	/// Sets the scissor test for future render calls. Future render calls will be batched with this scissor test.
-	/// </summary>
-	public static void SetScissor(Scissor scissor)
+	public static Scissor? GetCalculatedScissor()
 	{
-		CurrentScissor = scissor;
+		if (_activeScissors.Count == 0)
+			return null;
+
+		Scissor scissor = _activeScissors[0];
+
+		for (int i = 1; i < _activeScissors.Count; i++)
+		{
+			Scissor scissorToCombine = _activeScissors[i];
+
+			int x1 = Math.Max(scissor.X, scissorToCombine.X);
+			int y1 = Math.Max(scissor.Y, scissorToCombine.Y);
+
+			int aX2 = scissor.X + (int)scissor.Width;
+			int aY2 = scissor.Y + (int)scissor.Height;
+			int bX2 = scissorToCombine.X + (int)scissorToCombine.Width;
+			int bY2 = scissorToCombine.Y + (int)scissorToCombine.Height;
+
+			int x2 = Math.Min(aX2, bX2) - x1;
+			int y2 = Math.Min(aY2, bY2) - y1;
+
+			scissor = new(x1, y1, (uint)x2, (uint)y2);
+		}
+
+		return scissor;
 	}
 
 	/// <summary>
-	/// Unsets the scissor test for future render calls. Future render calls will not be batched with a scissor test.
+	/// Pushes a scissor test for future render calls. Future render calls will be batched with this scissor test.
 	/// </summary>
-	public static void UnsetScissor()
+	public static void PushScissor(Scissor scissor)
 	{
-		CurrentScissor = null;
+		_activeScissors.Add(scissor);
+	}
+
+	/// <summary>
+	/// Pops the most recent scissor test. Future render calls will not be batched with this scissor test.
+	/// </summary>
+	public static void PopScissor()
+	{
+		_activeScissors.RemoveAt(_activeScissors.Count - 1);
 	}
 }